excel_templates/upload_scripts/requirements.txt
pandas>=1.5.0
openpyxl>=3.0.10
pyodbc>=4.0.35
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
xlsxwriter>=3.0.0
colorama>=0.4.5
tqdm>=4.64.0 